Use FIOFFS on illumos instead of fsync
illumos supports sending a FIOFFS ioctl to a ZFS dataset in order to
sync _all_ outstanding IO. Use this ioctl in `region_flush` for the
whole region dataset instead of calling `sync_all` for each extent file.
This is hidden behind a new `omicron-build` feature, which is true for
our production builds.

This necessitated separating flush (which commits bits to durable
storage and is a no-op if `omicron-build` is used) and `post_flush`
(which performs clean up or any other kind of accounting). Splitting
this up exposed a bug in the sqlite implementation of ExtentInner: it
wasn't checking to see if the extent was dirty and wasn't calling
`set_flush_number` before calling `sync_all` on the user data file.